Create lengthOfLongestSubstring.py
Checks of the longest substring in a string by using sliding window and a dictionary.

The code checks how long the right windows can extend until there is a duplicate character and if there is a duplicate, moves the left window to the right until there is no duplicate. It repeats the step until right window extends to the length of the string.